requirements.txt
Flask==2.0.3
tensorflow==2.9.1
numpy==1.21.5
pyjwt==2.3.0
werkzeug==2.0.3
